apis/aws-bedrock-integrations/requirements.txt
fastapi==0.111.1
uvicorn==0.23.2
python-dotenv==1.0.0
openai==1.58.1
pymupdf==1.23.7
pyJWT==2.8.0
pydantic==2.8.2 
pymongo==4.6.2
boto3==1.35.97
botocore==1.35.97
